- [imp] account,bot: add different user levels of access as normal, admin & god. - Admin users now can upgrade/downgrade users premiumness, other than normal user stuff. - Add Add/Remove Admin commands and keyboard buttons for God users, so that they can introduce or remove staff as bot admins. - Also did some performance improvements, typo fixing and cleanup on BotMan and account classes.
